feat(invoice): per-line VAT on expense items (#99)
* feat(invoice): vat_rate on expense line items

Add optional vat_rate (Decimal percent, 0-100, default 0) to InvoiceItem
value object. total is now TTC (total_ht + total_tva). Persist and read
vat_rate in JSONB; legacy rows without the key default to 0.
sum_funds_released is now VAT-aware (qty × price × (1 + vat/100)).

The funds-release bridge copies vat_rate per line and no longer appends
synthetic "TVA {rate}%" lines; the expense TTC equals the facture TTC
through per-line math instead of bucket aggregation.

Thread vat_rate through API schemas, create/update use-cases, and DTOs.

* feat(invoice): restructure auto funds-release items with per-line vat_rate

Add migration a1c3e5f7b2d4 (down_revision b8e2d4f6a1c3). Regenerates
items of all auto-generated linked released_funds invoices from their
source billing document as {description, quantity, unit_price, vat_rate}
floats — no synthetic TVA aggregation lines.

Downgrade restores the b8e2d4f6a1c3 form (HT lines + per-rate TVA lines,
2 dp ROUND_HALF_UP, rate descending, zero buckets skipped). Idempotent
full regeneration from the source document; unlinked and manual rows
are untouched.

* fix(invoice): quantize money to cents at API serialization

Per-line VAT totals keep full Decimal precision in the domain; API
responses round item totals, total_amount, and funds_released_total to
2 dp HALF_UP so displayed expense amounts match billing document totals.

* chore: bump version to 0.2.79

---------

+++ app/api/v1/invoices/invoice_routes.py
@@ -23,6 +23,7 @@ from app.application.invoice import (
     ListInvoicesRequest,
     UpdateInvoiceRequest,
 )
+from app.application.invoice.dtos import money
 from app.domain.companies.exceptions import ForbiddenCompanyError
 from app.domain.entities.invoice import InvoiceType
 from app.domain.exceptions.invoice_exceptions import (
@@ -107,7 +108,7 @@ def list_invoices(project_id: str):
     except ValueError as e:
         return _error_response("ValidationError", str(e), 400)
 
-    funds_released_total = float(container.invoice_repository.sum_funds_released(UUID(project_id)))
+    funds_released_total = money(container.invoice_repository.sum_funds_released(UUID(project_id)))
 
     return jsonify(
         {
